// src/app/models/PlanItem.ts
export interface PlanItem {
  title: string;
  description: string;
  start: Date;
  end: Date;
  worker_title: string;
  worker: string;
  worker_cover?: any;
  lesson_form: string;
  lesson_form_short: string;
  group_name: string;
  tok_name: string;
  room: string;
  lesson_status: string;
  lesson_status_short: string;
  status_item: string;
  subject: string;
  hours: string;
  color: string;
  borderColor: string;
}